+++ OptimoveCore/Sources/Classes/NetworkClient/NetworkClient.swift
@@ -53,6 +53,10 @@ extension NetworkClientImpl: NetworkClient {
 
         request.headers?.forEach { urlRequest.addValue($0.value, forHTTPHeaderField: $0.field) }
 
+        // Signal that this SDK version supports auth.
+        urlRequest.addValue("1", forHTTPHeaderField: "X-Optimove-Auth-Capable")
+        urlRequest.addValue("ios", forHTTPHeaderField: "X-Optimove-Platform")
+
         let task = session.dataTask(with: urlRequest) { data, response, error in
             if let error = error {
                 completion(.failure(NetworkError.error(error)))
@@ -65,7 +69,9 @@ extension NetworkClientImpl: NetworkClient {
             switch httpResponse.statusCode {
             case 200 ... 299:
                 completion(.success(NetworkResponse<Data?>(statusCode: httpResponse.statusCode, body: data)))
-            case 400 ... 499:
+            case 401:
+                completion(.failure(NetworkError.unauthorized(data)))
+            case 400, 402 ... 499:
                 completion(.failure(NetworkError.requestInvalid(data)))
             case 500 ... 599:
                 completion(.failure(NetworkError.requestFailed))